# larapp/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SIMAS MTTG API",
        "description": "Backend API untuk Sistem Informasi Manajemen Masjid (SIMAS) Majelis Taklim Telkom Group.",
        "version": "1.0.0"
    },
    "paths": {
...
        "/mosques": {
            "get": {
                "tags": [
                    "Mosque"
                ],
                "summary": "List masjid/musholla",
                "description": "Ambil daftar masjid/musholla dengan filter pencarian dan pagination.",
                "operationId": "8bcd0755a4c54f063edcf3a21b65666e",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Cari berdasarkan nama/alamat",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "province_id",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "city_id",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "witel_id",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "type",
                        "in": "query",
                        "description": "Jenis: MASJID atau MUSHOLLA",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "MASJID",
                                "MUSHOLLA"
                            ]
                        }
                    },
                    {
                        "name": "min_completion",
                        "in": "query",
                        "description": "Minimal persentase kelengkapan fasilitas (0-100)",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "sort_by",
                        "in": "query",
                        "description": "Urutkan berdasarkan",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "name",
                                "completion",
                                "newest"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Daftar masjid/musholla",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "properties": {
                                                "items": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "integer"
                                                            },
                                                            "name": {
                                                                "type": "string"
                                                            },
                                                            "type": {
                                                                "type": "string"
                                                            },
                                                            "address": {
                                                                "type": "string",
                                                                "nullable": true
                                                            },
                                                            "province_id": {
                                                                "type": "integer",
                                                                "nullable": true
                                                            },
                                                            "city_id": {
                                                                "type": "integer",
                                                                "nullable": true
                                                            },
                                                            "witel_id": {
                                                                "type": "integer",
                                                                "nullable": true
                                                            },
                                                            "completion_percentage": {
                                                                "type": "integer"
                                                            },
                                                            "is_active": {
                                                                "type": "boolean"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                },
                                                "meta": {
                                                    "properties": {
                                                        "current_page": {
                                                            "type": "integer"
                                                        },
                                                        "per_page": {
                                                            "type": "integer"
                                                        },
                                                        "total": {
                                                            "type": "integer"
                                                        },
                                                        "last_page": {
                                                            "type": "integer"
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Mosques"
                ],
                "summary": "Tambah masjid/musholla",
                "operationId": "aa635d916fce3627b68d1e473096f771",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "type"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "code": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "type": {
                                        "type": "string",
                                        "enum": [
                                            "MASJID",
                                            "MUSHOLLA"
                                        ]
                                    },
                                    "address": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "province_id": {
                                        "type": "integer",
                                        "nullable": true
                                    },
                                    "city_id": {
                                        "type": "integer",
                                        "nullable": true
                                    },
                                    "witel_id": {
                                        "type": "integer",
                                        "nullable": true
                                    },
                                    "latitude": {
                                        "type": "number",
                                        "format": "float",
                                        "nullable": true
                                    },
                                    "longitude": {
                                        "type": "number",
                                        "format": "float",
                                        "nullable": true
                                    },
                                    "image_url": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "description": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "is_active": {
                                        "type": "boolean",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Data berhasil dibuat",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/mosques/{id}/facilities": {
            "put": {
                "tags": [
                    "Mosques",
                    "Facilities"
                ],
                "summary": "Update fasilitas masjid/musholla",
                "operationId": "c29bb3215c2fd8365c22800d18174190",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "facilities": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "facility_id": {
                                                    "type": "integer"
                                                },
                                                "is_available": {
                                                    "type": "boolean"
                                                },
                                                "note": {
                                                    "type": "string",
                                                    "nullable": true
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Fasilitas berhasil diperbarui",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}